restructured files, fixed navbar and footer position
