+++ client/src/App.jsx
@@ -1,22 +1,21 @@
 import React from 'react';
 // import { AppContextProvider } from './context/AppContext';
-import Homepage from './components/Homepage/Homepage';
-import LocalVoice from './components/LocalVoice/LocalVoice';
-import MsgRep from './components/MsgRep/MsgRep';
+import Homepage from './pages/Homepage/Homepage';
+import LocalVoice from './pages/LocalVoice/LocalVoice';
+import MsgRep from './pages/MsgRep/MsgRep';
 import Navbar from './components/Navbar';
-import WhoRepYou from './components/WhoRepYou/WhoRepYou';
-import YvYv from './components/YvYv/YvYv';
-import RepProfile from './components/RepProfile/RepProfile';
+import WhoRepYou from './pages/WhoRepYou/WhoRepYou';
+import YvYv from './pages/YvYv/YvYv';
+import RepProfile from './pages/RepProfile/RepProfile';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import Footer from './components/Footer';
-import FindMyRep from './components/FindMyRepPage/FindMyRep';
-import YourReps from './components/YourRepsPage/YourReps';
+import FindMyRep from './pages/FindMyRepPage/FindMyRep';
+import YourReps from './pages/YourRepsPage/YourReps';
 import { AppContextProvider } from './context/AppContext';
 import './App.css';
 
 const App = () => {
   return (
-
     <AppContextProvider>
       <BrowserRouter>
         <Navbar />
@@ -24,7 +23,7 @@ const App = () => {
           <Route exact path="/" component={Homepage} />
           <Route exact path="/LocalVoice" component={LocalVoice} />
           <Route exact path="/MsgRep" component={MsgRep} />
-          <Route exact path="/ThankYou" component={ThankYou} />
+          {/* <Route exact path="/ThankYou" component={ThankYou} /> */}
           <Route exact path="/WhoRepYou" component={WhoRepYou} />
           <Route exact path="/YvYv" component={YvYv} />
           <Route exact path="/RepProfile" component={RepProfile} />
@@ -34,7 +33,6 @@ const App = () => {
         <Footer />
       </BrowserRouter>
     </AppContextProvider>
-
   );
 };
